+++ react/components/StyledLink.tsx
@@ -59,7 +59,7 @@ const StyledLink: FunctionComponent<StyledLinkProps> = props => {
   )
 
   const iconTestId = `icon-${iconPosition}`
-  const iconComponent = iconProps ? (
+  const iconComponent = iconProps?.id ? (
     <span className={`${handles.styledLinkIcon} mh2`} data-testid={iconTestId}>
       <Icon
         id={iconProps.id}