+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Add prop 'isOpenOnMount' to allow opening submenu on mount
 
 ## [2.32.0] - 2021-06-24
+
 ### Added
 - Add CSS handles modifiers of open and close to menu item.