+++ CHANGELOG.md
@@ -6,10 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
 ### Added
-
 - Add prop 'onMountBehavior' to allow opening submenu on mount
+
 ## [2.32.0] - 2021-06-24
 
 ### Added