+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added 
+- I18n Bg and Bs
+
+### Fixed
+- I18n Fr, It, Ja, Ko, Nl and Ro
+
 ## [2.32.1] - 2021-09-13
 ### Added
 - Update schema to be able to change svg icon from Site Editor